requirements.txt
Django==4.1.3
asgiref==3.6.0
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7